refactor(new): Consolidate code
